+++ src/main/java/com/example/calculator/service/InputService.java
@@ -186,4 +186,22 @@ public class InputService {
         }
         return str.charAt(str.length() - 1);
     }
+
+    /**
+     * 桁数をチェックして、入力継続（指定桁数以内であれば）できるならtrueを返す
+     *
+     * <p>このメソッドは、文字列を数式として分解し、最後のセグメントが数字であればその桁数を数え、桁数が11桁以下であればtrueを返します。</p>
+     *
+     * @param formula チェック対象の文字列（数式として分解される）
+     * @return 最後のセグメントが数字で、かつその桁数が11桁以下であればtrue、それ以外はfalse
+     */
+    public boolean checkNumber(String formula) {
+        int count = 0;
+        String[] segments = Calculation.splitFormula(formula); // 分解してまとまりに分ける処理
+        String lastSegment = segments[segments.length - 1];
+        if (CommonUtil.isNum(lastSegment)) {
+            count = CommonUtil.countDigits(segments[segments.length - 1]);
+        }
+        return count <= 11;
+    }
 }
