+++ src/main/java/com/example/calculator/util/CommonUtil.java
@@ -26,24 +26,6 @@ public class CommonUtil {
     }
 
     /**
-     * 桁数をチェックして、入力継続（指定桁数以内であれば）できるならtrueを返す
-     *
-     * <p>このメソッドは、文字列を数式として分解し、最後のセグメントが数字であればその桁数を数え、桁数が11桁以下であればtrueを返します。</p>
-     *
-     * @param formula チェック対象の文字列（数式として分解される）
-     * @return 最後のセグメントが数字で、かつその桁数が11桁以下であればtrue、それ以外はfalse
-     */
-    public static boolean checkNumber(String formula) {
-        int count = 0;
-        String[] segments = Calculation.splitFormula(formula); // 分解してまとまりに分ける処理
-        String lastSegment = segments[segments.length - 1];
-        if (isNum(lastSegment)) {
-            count = countDigits(segments[segments.length - 1]);
-        }
-        return count <= 11;
-    }
-
-    /**
      * 与えられた文字列が数字または小数点を含む数字であるかをチェックします。
      *
      * <p>このメソッドは、文字列が以下のいずれかの形式である場合にtrueを返します：</p>
